+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added details prop to SplashError
 
+- Added LabeledInput component
+
 ## [0.15.1] - 2020-04-09
 
 - Loading Button state looks more disabled than before